build: handle do_deploy refactoring
With the removal of IMAGE_POSTPROCESS_COMMAND for
xenclient-installer-image, the ISO and PXE files end up in slightly
different directories.
The postprocess commands were able to use IMAGE_LINK_NAME to create a
subdirectory, which is not possible in do_deploy.

Still, check for the former directories in case this version of
xenclient-oe does not have the refactoring yet.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>